% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_program.R
\name{ui_program}
\alias{ui_program}
\title{shiny UI element}
\usage{
ui_program(type = "tab", title = "Breeding program configuration",
  name = "resource_program", output = "hot_program")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

